--- packages/race-cli/src/reporters.ts
@@ -1,141 +1,0 @@
-import fs from 'fs/promises';
-import {
-  LighthouseResultsWrapper,
-  LighthouseResults,
-  LightHouseAuditKeys,
-} from '@racepoint/shared';
-import {
-  connectRepository,
-  ReportingRepository,
-  ReportingRow,
-} from './repository';
-import logger from './logger';
-
-export interface LLReporter {
-  initialize: () => Promise<void>;
-  process: (results: LighthouseResultsWrapper) => Promise<void> | undefined;
-}
-
-interface MetricHeader {
-  name: string;
-  key: string;
-}
-
-export class ConsoleReporter implements LLReporter {
-  private metricHeaders: MetricHeader[] = [
-    {name: 'SI', key: LightHouseAuditKeys.SI},
-    {name: 'FCP', key: LightHouseAuditKeys.FCP},
-    {name: 'LCP', key: LightHouseAuditKeys.LCP},
-    {name: 'CLS', key: LightHouseAuditKeys.CLS},
-    {name: 'MP-FID', key: LightHouseAuditKeys.MaxFID},
-    {name: 'TBT', key: LightHouseAuditKeys.TotalBlocking},
-  ];
-
-  // state to know if we've begun writing to the screen yet
-  private _hasBegun: boolean;
-
-  constructor() {
-    this._hasBegun = false;
-  }
-
-  initialize = (): Promise<void> => Promise.resolve();
-
-  process = (results: LighthouseResultsWrapper): Promise<void> =>
-    new Promise((resolve) => {
-      if (this._hasBegun === false) {
-        this._hasBegun = true;
-        console.log(this.buildHeaders());
-      }
-      console.log(this.buildRow(results.lhr));
-      resolve();
-    });
-
-  private _buildMetricsRow = (
-    mapper: (header: MetricHeader) => string
-  ): string => this.metricHeaders.map((header) => mapper(header)).join('\t|\t');
-
-  private buildHeaders = (): string =>
-    this._buildMetricsRow((header) => header.name);
-
-  private buildRow = (results: LighthouseResults): string =>
-    this._buildMetricsRow((header) => {
-      return results.audits[header.key].displayValue;
-    });
-}
-
-/**
- * Report Lighthouse results to some ReportingRepository.
- *
- * If repository id does not exist, will attempt to create one
- *
- */
-export class RepositoryReporter implements LLReporter {
-  private _repositoryLocation: string;
-  private _targetUrl: string;
-  private _repository: ReportingRepository | undefined; // lateinit
-
-  constructor(
-    targetUrl: string,
-    repositoryId: string | undefined,
-    outputTarget: string | undefined
-  ) {
-    this._targetUrl = targetUrl;
-    this._repositoryLocation = `${outputTarget}/${repositoryId}`;
-  }
-
-  initialize = (): Promise<void> =>
-    connectRepository(this._repositoryLocation)
-      .then((receivedRepository: ReportingRepository) => {
-        this._repository = receivedRepository;
-      })
-      .catch((e) => {
-        logger.error('Failed to connect to repository');
-      });
-
-  process = (results: LighthouseResultsWrapper): Promise<void> | undefined => {
-    return (
-      this._repository &&
-      this._repository.write(this.mapResultsToRow(results.lhr))
-    );
-  };
-
-  private mapResultsToRow = (results: LighthouseResults): ReportingRow => {
-    const row = {
-      dateOccured: new Date().toISOString(),
-      profiledUrl: this._targetUrl,
-      lighthouseTime: results.timing.total,
-      speedIndex: results.audits[LightHouseAuditKeys.SI].numericValue,
-      fcp: results.audits[LightHouseAuditKeys.FCP].numericValue,
-      lcp: results.audits[LightHouseAuditKeys.LCP].numericValue,
-      cls: results.audits[LightHouseAuditKeys.CLS].numericValue,
-      mpfid: results.audits[LightHouseAuditKeys.MaxFID].numericValue,
-      tbt: results.audits[LightHouseAuditKeys.TotalBlocking].numericValue,
-    };
-    return row;
-  };
-}
-
-/**
- * Writes a single lighthouse result to an html file
- */
-export class HtmlReporter implements LLReporter {
-  private _reportPath: string;
-  constructor(outputTarget: string) {
-    this._reportPath = outputTarget; //`${outputTarget}/results.html`;
-  }
-
-  // for now, just resolve. We can get fancy and not overwrite by checking fs.access and selecting a new reportPath name
-  initialize = (): Promise<void> => Promise.resolve();
-
-  process = (results: LighthouseResultsWrapper): Promise<void> => {
-    const htmlPath = `${this._reportPath}/results_${results.lhr.fetchTime}.html`;
-    return fs
-      .writeFile(htmlPath, results.report, {flag: 'w'})
-      .then(() => {
-        logger.info(`Lighthouse HTML results successfully saved`);
-      })
-      .catch((e) => {
-        console.log('Failed to write results', e);
-      });
-  };
-}
