--- packages/race-cli/src/scenarios.ts
@@ -1,4 +1,0 @@
-import {Scenario} from './types';
-import {ProfileScenario} from './profile';
-
-export const scenarios: Scenario<any>[] = [new ProfileScenario()];
